+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - `IFS-334`: Hinzufügen eines Profils für Jacoco und Anpassung der Surefire Config
 - `IFS-410`: Checkstyle Plugin auf Version 3.1.0 angehoben
 - `IFS-454`: Maven Enforcer Rule: dependency-convergence ergänzt.
+- `IFS-436`: Entkopplung Health-Endpoint und hinzufügen Nachbarsystem-Check
 - `IFS-458`: Zulieferung von Github getestet und übernommen. (Versionsanhebungen rausgenommen.)
 - `IFS-463`: Abwärtskompatibilität IsyFact 2 zu IsyFact 1 - Erstellung der Bridge
 - `IFS-466`: Fix für fehlerhafte Mdc Formatierung